refactor: moved currentDate to utils.ts
